+++ MANIFEST.in
@@ -1,2 +1,4 @@
-include *.md
 include LICENSE
+include README.rst *.rst *.md
+recursive-exclude * __pycache__
+recursive-exclude * *.py[co]
