--- drf-nested-routers/MANIFEST.in
@@ -1,4 +1,0 @@
-include LICENSE
-include README.rst *.rst *.md
-recursive-exclude * __pycache__
-recursive-exclude * *.py[co]
