--- drf-nested-routers/rest_framework_nested/__init__.py
@@ -1,1 +1,0 @@
-__version__ = '0.12.0'
