--- drf-nested-routers/setup.cfg
@@ -1,5 +1,0 @@
-[metadata]
-description-file = README.md
-
-[wheel]
-universal = 1
