+++ setup.cfg
@@ -1,2 +1,5 @@
 [metadata]
 description-file = README.md
+
+[wheel]
+universal = 1
