+++ setup.py
@@ -1,20 +1,94 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+import re
 import os
+import sys
+from setuptools import setup
 
-from setuptools import find_packages, setup
 
-curdir = os.path.dirname(os.path.abspath(__file__))
+name = 'drf-nested-routers'
+package = 'rest_framework_nested'
+description = 'Nested resources for the Django Rest Framework'
+url = 'https://github.com/alanjds/drf-nested-routers'
+author = 'Alan Justino et al.'
+author_email = 'alan.justino@yahoo.com.br'
+license = 'Apache'
+
+
+def get_version(package):
+    """
+    Return package version as listed in `__version__` in `init.py`.
+    """
+    init_py = open(os.path.join(package, '__init__.py')).read()
+    return re.search("^__version__ = ['\"]([^'\"]+)['\"]",
+                     init_py, re.MULTILINE).group(1)
+
+
+def get_packages(package):
+    """
+    Return root package and all sub-packages.
+    """
+    return [dirpath
+            for dirpath, dirnames, filenames in os.walk(package)
+            if os.path.exists(os.path.join(dirpath, '__init__.py'))]
+
+
+def get_package_data(package):
+    """
+    Return all files under the root package, that are not in a
+    package themselves.
+    """
+    walk = [(dirpath.replace(package + os.sep, '', 1), filenames)
+            for dirpath, dirnames, filenames in os.walk(package)
+            if not os.path.exists(os.path.join(dirpath, '__init__.py'))]
+
+    filepaths = []
+    for base, filenames in walk:
+        filepaths.extend([os.path.join(base, filename)
+                          for filename in filenames])
+    return {package: filepaths}
+
+
+version = get_version(package)
+
+
+if sys.argv[-1] == 'publish':
+    if os.system("pip freeze | grep wheel"):
+        print("wheel not installed.\nUse `pip install wheel`.\nExiting.")
+        sys.exit()
+    os.system("python setup.py sdist upload")
+    os.system("python setup.py bdist_wheel upload")
+    print("You probably want to also tag the version now:")
+    print("  git tag -a {0} -m 'version {0}'".format(version))
+    print("  git push --tags")
+    sys.exit()
+
 
 setup(
-    name='drf-nested-routers',
-    description='Nested resources for the Django Rest Framework',
+    name=name,
+    version=version,
+    url=url,
+    license=license,
+    description=description,
     long_description=open('README.md').read(),
-    license='Apache',
-    version='0.11.0',
-    author='Alan Justino et al.',
-    author_email='alan.justino@yahoo.com.br',
-    url='https://github.com/alanjds/drf-nested-routers',
+    author=author,
+    author_email=author_email,
+    packages=get_packages(package),
+    package_data=get_package_data(package),
     install_requires=['djangorestframework>=2.4'],
-    setup_requires=['setuptools'],
-    packages=find_packages(curdir),
+    classifiers=[
+        'Development Status :: 5 - Production/Stable',
+        'Environment :: Web Environment',
+        'Framework :: Django',
+        'Intended Audience :: Developers',
+        'License :: OSI Approved :: Apache Software License',
+        'Operating System :: OS Independent',
+        'Natural Language :: English',
+        'Programming Language :: Python :: 2',
+        'Programming Language :: Python :: 2.7',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.3',
+        'Programming Language :: Python :: 3.4',
+        'Topic :: Internet :: WWW/HTTP',
+    ]
 )
